requirements.txt
Django==1.11.4
django-analytical==2.2.2
django-compat==1.0.14
django-extensions==1.7.5
django-ipware==1.1.6
django-leaflet==0.22.0
django-geojson==2.10.0
django-pipeline==1.6.10
elasticsearch==2.3.0
geoip2==0.5.0
google-api-python-client==1.6.2
MySQL-python==1.2.5
oauth2==1.5.211
oauth2client==4.1.1
oauthlib==1.0.3
progressbar==2.3
psycopg2==2.7.3
python-memcached==1.53
pytz==2017.2
virtualenv==12.1.1
